requirements.txt
python-dotenv
pymupdf
requests
beautifulsoup4
openai
langchain
langchain-openai
Flask
colorama
langchain-community
chromadb